Improve mysql confirmation code cleanup script

+++ features/identity-core/org.wso2.carbon.identity.core.server.feature/resources/dbscripts/stored-procedures/mysql/confirmation-code-cleanup/mysql-confirmation-code-cleanup.sql
@@ -13,6 +13,9 @@ BEGIN
     DECLARE rowCount INT;
     DECLARE enableLog BOOLEAN;
     DECLARE backupTables BOOLEAN;
+    DECLARE sleepTime FLOAT;
+    DECLARE cleanUpCodesTimeLimit INT;
+    DECLARE cleanUpDateTimeLimit DATETIME;
 
     -- ------------------------------------------
     -- CONFIGURE VARIABLES
@@ -20,19 +23,30 @@ BEGIN
     SET batchSize    = 10000; -- SET BATCH SIZE FOR AVOID TABLE LOCKS [DEFAULT : 10000]
     SET chunkSize    = 500000; -- CHUNK WISE DELETE FOR LARGE TABLES [DEFAULT : 500000]
     SET enableLog    = FALSE; -- ENABLE LOGGING [DEFAULT : FALSE]
-    SET backupTables = FALSE; -- SET IF TOKEN TABLE NEEDS TO BACKUP BEFORE DELETE [DEFAULT : FALSE] , WILL DROP THE PREVIOUS BACKUP TABLES IN NEXT ITERATION
+    SET backupTables = FALSE; -- SET IF RECOVERY DATA TABLE NEEDS TO BACKED-UP BEFORE DELETE [DEFAULT : FALSE] , WILL DROP THE PREVIOUS BACKUP TABLES IN NEXT ITERATION
+    SET sleepTime = 2; -- SET SLEEP TIME FOR AVOID TABLE LOCKS     [DEFAULT : 2]
+    SET cleanUpCodesTimeLimit = 720;  -- SET SAFE PERIOD OF HOURS FOR CODE DELETE [DEFAULT : 720 hrs (30 days)]. CODES OLDER THAN THE NUMBER OF HOURS DEFINED HERE WILL BE DELETED.
 
     SET batchCount = 1000;
     SET chunkCount = 1000;
     SET rowCount   = 0;
+    SET cleanUpDateTimeLimit = DATE_ADD(NOW(), INTERVAL -(cleanUpCodesTimeLimit) HOUR);
 
-    SELECT 'WSO2_CONFIRMATION_CODE_CLEANUP() STARTED...!' AS 'INFO LOG';
+    IF (enableLog)
+    THEN
+        SELECT 'WSO2_CONFIRMATION_CODE_CLEANUP() STARTED...!' AS 'INFO LOG';
+    END IF;
 
     -- ------------------------------------------
     -- BACKUP DATA
     -- ------------------------------------------
     IF (backupTables)
     THEN
+        IF (enableLog)
+        THEN
+            SELECT 'TABLE BACKUP STARTED ... !' AS 'INFO LOG';
+        END IF;
+
         IF (EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = 'BAK_IDN_RECOVERY_DATA' and TABLE_SCHEMA in (SELECT DATABASE())))
         THEN
             IF (enableLog)
@@ -48,26 +62,38 @@ BEGIN
     -- ------------------------------------------
     -- CLEANUP DATA
     -- ------------------------------------------
-    WHILE (chunkCount > 0) DO
+    CONFIRMATION_CHUNK_LOOP : WHILE (chunkCount > 0) DO
         -- CREATE CHUNK TABLE
         DROP TABLE IF EXISTS IDN_RECOVERY_DATA_CHUNK_TMP;
-        CREATE TABLE IDN_RECOVERY_DATA_CHUNK_TMP AS SELECT CODE FROM IDN_RECOVERY_DATA LIMIT chunkSize;
+        CREATE TABLE IDN_RECOVERY_DATA_CHUNK_TMP AS SELECT CODE FROM IDN_RECOVERY_DATA WHERE (cleanUpDateTimeLimit > TIME_CREATED) LIMIT chunkSize;
         SELECT row_count() INTO chunkCount;
         CREATE INDEX IDN_RECOVERY_DATA_CHUNK_TMP_INDX on IDN_RECOVERY_DATA_CHUNK_TMP (CODE);
         COMMIT;
+
+        IF (chunkCount = 0)
+        THEN
+            LEAVE CONFIRMATION_CHUNK_LOOP;
+        END IF;
+
         IF (enableLog) THEN
             SELECT 'CREATED IDN_RECOVERY_DATA_CHUNK_TMP...' AS 'INFO LOG';
         END IF;
 
         -- BATCH LOOP
         SET batchCount = 1;
-        WHILE (batchCount > 0) DO
+        CONFIRMATION_BATCH_LOOP : WHILE (batchCount > 0) DO
             -- CREATE BATCH TABLE
             DROP TABLE IF EXISTS IDN_RECOVERY_DATA_BATCH_TMP;
             CREATE TABLE IDN_RECOVERY_DATA_BATCH_TMP AS
               SELECT CODE FROM IDN_RECOVERY_DATA_CHUNK_TMP LIMIT batchSize;
             SELECT row_count() INTO batchCount;
             COMMIT;
+
+            IF (batchCount = 0)
+            THEN
+                LEAVE CONFIRMATION_BATCH_LOOP;
+            END IF;
+
             IF (enableLog) THEN
                 SELECT 'CREATED IDN_RECOVERY_DATA_BATCH_TMP...' AS 'INFO LOG';
             END IF;
@@ -85,7 +111,7 @@ BEGIN
             COMMIT;
 
             IF (enableLog) THEN
-                SELECT 'BATCH DELETE FINISHED ON IDN_RECOVERY_DATA : ' + rowCount AS 'INFO LOG';
+                SELECT 'BATCH DELETE FINISHED ON IDN_RECOVERY_DATA : ' AS 'INFO LOG', rowCount;
             END IF;
 
             -- DELETE FROM CHUNK
@@ -93,6 +119,11 @@ BEGIN
               FROM IDN_RECOVERY_DATA_CHUNK_TMP AS A
               INNER JOIN IDN_RECOVERY_DATA_BATCH_TMP AS B
               ON A.CODE = B.CODE;
+
+            IF ((rowCount > 0))
+            THEN
+                DO SLEEP(sleepTime);
+            END IF;
 
         END WHILE;
     END WHILE;
